Unify span implementation for gather and gather_nd.
Create span.hpp for common implementation of span.
